requirements.txt
flask
flask-socketio
eventlet
flask-cors
pytest
requests
websocket-client
selenium
Appium-Python-Client